requirement.txt
# Requirements for the main script
SpeechRecognition==3.8.1
pygame==2.1.2
requests==2.31.0
beautifulsoup4==4.11.2
pyautogui==0.9.53
characterai==0.1.4
speedtest-cli==2.1.3

# Requirements for the GUI script
numpy==1.24.2
opencv-python==4.6.0.66
Pillow==9.5.0
pydub==0.25.1

# Note: Google Generative AI and specific imports from 'google.generativeai'
# are not standard and may need specific handling. Ensure you have the necessary SDKs or packages for 'google.generativeai'.
